requirements.txt
google-genai
python-dotenv
jiwer
numpy
tqdm
httpx
fastapi

 